+++ packages/common/src/types/ContractName.ts
@@ -82,6 +82,15 @@ export function toSchemaValue(contractName: ContractName): SchemaValue {
 }
 
 /**
+ * Convert a contract name format used by schema.
+ * @param {SchemaValue} contractName The contract name in schema format.
+ * @returns {ContractName} The contract name.
+ */
+export function fromSchemaValue(contractName: SchemaValue): ContractName {
+    return fromString(contractName.contract);
+}
+
+/**
  * Check if two contract names represent the same name of a contract.
  * @param {ContractName} left
  * @param {ContractName} right