+++ packages/common/test/types/blockItemSummary.test.ts
@@ -27,6 +27,7 @@ import {
     InitName,
     Parameter,
     ReceiveName,
+    ContractEvent,
 } from '../../src/index.js';
 
 const chainUpdate: UpdateSummary = {
@@ -106,14 +107,14 @@ const contractUpdate: UpdateContractSummary & BaseAccountTransactionSummary = {
             ),
             events: [
                 'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-            ],
+            ].map(ContractEvent.fromHexString),
         },
         {
             tag: TransactionEventTag.Interrupted,
             address: ContractAddress.create(3496),
             events: [
                 'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-            ],
+            ].map(ContractEvent.fromHexString),
         },
         {
             tag: TransactionEventTag.Updated,
@@ -134,7 +135,7 @@ const contractUpdate: UpdateContractSummary & BaseAccountTransactionSummary = {
             ),
             events: [
                 'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-            ],
+            ].map(ContractEvent.fromHexString),
         },
         {
             tag: TransactionEventTag.Transferred,
@@ -393,24 +394,24 @@ describe('getSummaryContractUpdateLogs', () => {
 
     test('Returns list of unique account addresses corresponding to transaction', () => {
         const logs = getSummaryContractUpdateLogs(contractUpdate);
-        expect(logs).toEqual([
+        expectToEqual(logs, [
             {
                 address: ContractAddress.create(3496),
                 events: [
                     'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-                ],
+                ].map(ContractEvent.fromHexString),
             },
             {
                 address: ContractAddress.create(3496),
                 events: [
                     'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-                ],
+                ].map(ContractEvent.fromHexString),
             },
             {
                 address: ContractAddress.create(4416),
                 events: [
                     'ff006400c8d4bb7106a96bfa6f069438270bf9748049c24798b13b08f88fc2f46afb435f0087e3bec61b8db2fb7389b57d2be4f7dd95d1088dfeb6ef7352c13d2b2d27bb49',
-                ],
+                ].map(ContractEvent.fromHexString),
             },
         ]);
     });